Step2_Training_MIL/requirements.txt
numpy==1.22.3
pandas==1.3.5
Pillow==9.1.0
pytorch_lightning==1.6.2
scikit_image==0.19.2
scikit_learn==1.0.2
skimage==0.0
torch==1.10.2
torchvision==0.12.0
tqdm==4.63.1
